tests-requirements.txt
scipy
imread
nose
pillow
